Update bin2txt converter
