+++ tools/bin2txt/syscall.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python3
 #
-# Copyright (c) 2017, Intel Corporation
+# Copyright 2017, Intel Corporation
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -14,7 +14,7 @@
 #       the documentation and/or other materials provided with the
 #       distribution.
 #
-#     * Neither the name of Intel Corporation nor the names of its
+#     * Neither the name of the copyright holder nor the names of its
 #       contributors may be used to endorse or promote products derived
 #       from this software without specific prior written permission.
 #
@@ -50,19 +50,19 @@ STATE_IN_ENTRY = 1
 STATE_ENTRY_COMPLETED = 2
 STATE_COMPLETED = 3
 STATE_CORRUPTED_ENTRY = 4
-STATE_UNKNOWN_EVENT = 5
 
 CNT_NONE = 0
 CNT_ENTRY = 1
 CNT_EXIT = 2
 
+RESULT_SUPPORTED = 0
+
 CHECK_OK = 0
 CHECK_NO_ENTRY = 1
 CHECK_NO_EXIT = 2
 CHECK_WRONG_ID = 3
-CHECK_WRONG_EXIT = 4
-CHECK_SAVE_IN_ENTRY = 5
-CHECK_NOT_FIRST_PACKET = 6
+CHECK_SAVE_IN_ENTRY = 4
+CHECK_NOT_FIRST_PACKET = 5
 
 EM_no_ret = 1 << 20  # syscall does not return
 
@@ -84,8 +84,8 @@ def is_exit(etype):
 # Syscall
 ########################################################################################################################
 class Syscall(SyscallInfo):
-    __str = "------------------ ------------------"
-    __arg_str_mask = [1, 2, 4, 8, 16, 32]
+    str_entry = "------------------ ------------------"
+    arg_str_mask = [1, 2, 4, 8, 16, 32]
 
     ####################################################################################################################
     def __init__(self, pid_tid, sc_id, sc_info, buf_size, debug_mode):
@@ -122,7 +122,6 @@ class Syscall(SyscallInfo):
         self.truncated = 0
 
         self.strings = []
-        self.str_is_path = []
 
         self.buf_size = int(buf_size)
         self.buf_size_2 = int(buf_size / 2)
@@ -138,13 +137,17 @@ class Syscall(SyscallInfo):
         self.fmt_exit = 'q'
         self.size_fmt_exit = struct.calcsize(self.fmt_exit)
 
+        self.pid_ind = -1
+        self.is_pmem = 0
+        self.unsupported_type = RESULT_SUPPORTED
+
     ####################################################################################################################
     def __lt__(self, other):
         return self.time_start < other.time_start
 
     ####################################################################################################################
     def is_string(self, n):
-        if (self.mask & self.__arg_str_mask[n]) == self.__arg_str_mask[n]:
+        if (self.mask & self.arg_str_mask[n]) == self.arg_str_mask[n]:
             return 1
         else:
             return 0
@@ -157,11 +160,7 @@ class Syscall(SyscallInfo):
         string = ""
         max_len = 0
 
-        if self.info_all >> 2:
-            max_len = self.str_max_1
-            string = buf_str
-
-        elif self.nstrargs == 1:
+        if self.nstrargs == 1 or self.info_all >> 2:
             max_len = self.str_max_1
             string = buf_str
 
@@ -172,9 +171,9 @@ class Syscall(SyscallInfo):
                 string = buf_str[0:self.buf_size_2]
             elif self.num_str == 2:
                 string = buf_str[self.buf_size_2: 2 * self.buf_size_2]
-            else:
+            else:  # pragma: no cover
                 assert_msg(self.num_str <= 2, "unsupported number of string arguments ({0:d}), "
-                                              "input file can be corrupted".format(self.nstrargs))
+                                              "input file may be corrupted".format(self.nstrargs))
 
         elif self.nstrargs == 3:
             max_len = self.str_max_3
@@ -185,12 +184,12 @@ class Syscall(SyscallInfo):
                 string = buf_str[self.buf_size_3: 2 * self.buf_size_3]
             elif self.num_str == 3:
                 string = buf_str[2 * self.buf_size_3: 3 * self.buf_size_3]
-            else:
+            else:  # pragma: no cover
                 assert_msg(self.num_str <= 3, "unsupported number of string arguments ({0:d}), "
-                                              "input file can be corrupted".format(self.nstrargs))
+                                              "input file may be corrupted".format(self.nstrargs))
 
-        else:
-            assert_msg(self.nstrargs <= 3, "unsupported number of string arguments ({0:d}), input file can be corrupted"
+        else:  # pragma: no cover
+            assert_msg(self.nstrargs <= 3, "unsupported number of string arguments ({0:d}), input file may be corrupted"
                        .format(self.nstrargs))
 
         str_p = str(string.decode(errors="ignore"))
@@ -219,11 +218,6 @@ class Syscall(SyscallInfo):
 
     ####################################################################################################################
     def print_single_record(self, debug):
-        if self.state not in (STATE_ENTRY_COMPLETED, STATE_COMPLETED):
-            if self.debug_mode and self.state != STATE_IN_ENTRY:
-                self.log_parse.debug("DEBUG(print_single_record): state = {0:d}".format(self.state))
-            return
-
         if self.state == STATE_ENTRY_COMPLETED:
             self.print_entry(debug)
             return
@@ -235,23 +229,24 @@ class Syscall(SyscallInfo):
                 self.print_exit(debug)
             return
 
+        if self.state == STATE_CORRUPTED_ENTRY:
+            self.log_print("corrupted entry packet information of syscall {0:s}:".format(self.name), debug)
+            self.log_print("0x{0:016X} 0x{1:016X} {2:s} {3:s} [corrupted entry packet]"
+                           .format(self.time_start, self.pid_tid, self.str_entry, self.name), debug)
+
     ####################################################################################################################
     def print_always(self):
-        if self.debug_mode and self.state not in (STATE_ENTRY_COMPLETED, STATE_COMPLETED):
-            self.log_parse.debug("DEBUG(print_always): state = {0:d}".format(self.state))
+        if self.debug_mode and self.state not in (STATE_INIT, STATE_ENTRY_COMPLETED, STATE_COMPLETED):
+            print("DEBUG(print_always): syscall: {0:s}, state: {1:d}".format(self.name, self.state))
 
         self.print_entry(DEBUG_OFF)
 
-        if (self.state == STATE_COMPLETED) and (self.mask & EM_no_ret == 0):
+        if (self.content & CNT_EXIT) and (self.mask & EM_no_ret == 0):
             self.print_exit(DEBUG_OFF)
 
-    ####################################################################################################################
-    # print -- print record only in non-debug mode
-    ####################################################################################################################
-    def print(self):
-        if self.debug_mode:
-            return
-        self.print_always()
+        if not (self.content & (CNT_ENTRY | CNT_EXIT)):
+            print("0x{0:016X} 0x{1:016X} {2:s} {3:s} [corrupted packet]".
+                  format(self.time_start, self.pid_tid, self.str_entry, self.name))
 
     ####################################################################################################################
     def log_print(self, msg, debug):
@@ -274,9 +269,9 @@ class Syscall(SyscallInfo):
             else:
                 print("WARNING: " + warn_str)
 
-        msg = "0x{0:016X} 0x{1:016X} {2:s} {3:s}".format(self.time_start, self.pid_tid, self.__str, self.name)
+        msg = "0x{0:016X} 0x{1:016X} {2:s} {3:s}".format(self.time_start, self.pid_tid, self.str_entry, self.name)
 
-        assert_msg(self.nargs == len(self.args), "incorrect number of syscall arguments, input file can be corrupted")
+        assert_msg(self.nargs == len(self.args), "incorrect number of syscall arguments, input file may be corrupted")
 
         for n in range(0, self.nargs):
             if self.is_string(n):
@@ -290,8 +285,7 @@ class Syscall(SyscallInfo):
     # print_exit -- print exit info of the syscall
     ####################################################################################################################
     def print_exit(self, debug):
-        if not (self.content & CNT_EXIT):
-            return
+        assert_msg(self.content & CNT_EXIT, "print_exit: no exit content")
 
         if len(self.name) > 0:
             self.log_print("0x{0:016X} 0x{1:016X} 0x{2:016X} 0x{3:016X} {4:s}".format(
@@ -321,10 +315,7 @@ class Syscall(SyscallInfo):
             return CHECK_NOT_FIRST_PACKET
 
         if self.state == STATE_INIT and is_exit(etype):
-            if sc_id == RT_SIGRETURN_SYS_EXIT:  # sys_exit of rt_sigreturn
-                return CHECK_OK
-
-            if retval == 0 and name in ("clone", "fork", "vfork"):
+            if sc_id == RT_SIGRETURN_SYS_EXIT or (retval == 0 and name in ("clone", "fork", "vfork")):
                 return CHECK_OK
 
             return CHECK_NO_ENTRY
@@ -336,23 +327,10 @@ class Syscall(SyscallInfo):
                 self.print_mismatch_info(etype, pid_tid, sc_id, name)
             return CHECK_SAVE_IN_ENTRY
 
-        if pid_tid == self.pid_tid and sc_id == self.sc_id:
-            wrong_id = 0
-        else:
-            wrong_id = 1
+        wrong_id = not (pid_tid == self.pid_tid and sc_id == self.sc_id)
 
-        if self.state == STATE_ENTRY_COMPLETED:
-            if debug_on and self.debug_mode and (is_entry(etype) or (is_exit(etype) and wrong_id)):
-                self.log_parse.debug("Notice: no exit info found for syscall: {0:016X} {1:s}"
-                                     .format(self.pid_tid, self.name))
-            if is_entry(etype):
-                return CHECK_NO_EXIT
-
-            if is_exit(etype) and wrong_id:
-                if retval == 0 and sc_id == self.sc_id and self.name in ("clone", "fork", "vfork"):
-                    # clone/fork/vfork returned 0 first
-                    return CHECK_NO_EXIT
-                return CHECK_WRONG_EXIT
+        if self.state == STATE_ENTRY_COMPLETED and (is_entry(etype) or (is_exit(etype) and wrong_id)):
+            return CHECK_NO_EXIT
 
         if wrong_id:
             if debug_on:
@@ -370,17 +348,16 @@ class Syscall(SyscallInfo):
         etype = info_all & E_MASK
         info_all &= ~E_MASK
 
+        assert_msg(etype in (E_KP_ENTRY, E_KP_EXIT, E_TP_EXIT), "unknown entry type")
+
         if etype == E_KP_ENTRY:
             if self.state not in (STATE_INIT, STATE_IN_ENTRY):
                 self.log_parse.error("wrong state for KProbe entry type: {0:d}".format(self.state))
             # kprobe entry handler
             return self.add_kprobe_entry(info_all, bdata, timestamp)
 
-        if (etype == E_KP_EXIT) or (etype == E_TP_EXIT):
-            # kprobe exit handler or raw tracepoint sys_exit
-            return self.add_exit(bdata, timestamp)
-
-        return STATE_UNKNOWN_EVENT
+        # kprobe exit handler or raw tracepoint sys_exit ((etype == E_KP_EXIT) or (etype == E_TP_EXIT))
+        return self.add_exit(bdata, timestamp)
 
     ####################################################################################################################
     # add_kprobe_entry -- add the kprobe entry info to the syscall record
@@ -410,7 +387,7 @@ class Syscall(SyscallInfo):
 
         # check if it is a continuation of a string
         if self.arg_first == self.arg_last:
-            assert_msg(self.is_cont == 1, "packet is not marked as a continuation, input file can be corrupted")
+            assert_msg(self.is_cont == 1, "packet is not marked as a continuation, input file may be corrupted")
 
             if self.str_fini:
                 return self.state
@@ -483,17 +460,10 @@ class Syscall(SyscallInfo):
         for n in range((self.arg_first - 1), self.arg_last):
             if self.is_string(n):
                 index = self.get_str_arg(buf_str)
-
                 if index >= 0:
-                    if len(self.args) < n + 1:
-                        self.args.append(index)
-                    else:
-                        self.args[n] = index
+                    self.args.append(index)
             else:
-                if len(self.args) < n + 1:
-                    self.args.append(args[n])
-                else:
-                    self.args[n] = args[n]
+                self.args.append(args[n])
 
         if end_of_syscall:
             self.num_str = 0  # reset counter of string arguments
@@ -511,21 +481,18 @@ class Syscall(SyscallInfo):
 
     ####################################################################################################################
     def get_return_value(self, bdata):
-        retval = -1
-        if len(bdata) >= self.size_fmt_exit:
-            bret = bdata[0: self.size_fmt_exit]
-            retval, = struct.unpack(self.fmt_exit, bret)
+        assert_msg(len(bdata) >= self.size_fmt_exit, "no return value, input file may be corrupted")
+        bret = bdata[0: self.size_fmt_exit]
+        retval, = struct.unpack(self.fmt_exit, bret)
         return retval
 
     ####################################################################################################################
-    # add_exit -- add the exit info to the syscall record
+    # save_exit -- save the exit info to the syscall record
     ####################################################################################################################
-    def add_exit(self, bdata, timestamp):
+    def save_exit(self, retval, timestamp):
         if self.state == STATE_INIT:
             self.time_start = timestamp
         self.time_end = timestamp
-
-        retval = self.get_return_value(bdata)
 
         # split return value into result and errno
         if retval >= 0:
@@ -538,6 +505,22 @@ class Syscall(SyscallInfo):
             self.err = -retval
 
         self.content |= CNT_EXIT
-        self.state = STATE_COMPLETED
 
+    ####################################################################################################################
+    # add_exit -- add the exit info to the syscall record and mark as completed
+    ####################################################################################################################
+    def add_exit(self, bdata, timestamp):
+        retval = self.get_return_value(bdata)
+        self.save_exit(retval, timestamp)
+        self.state = STATE_COMPLETED
         return self.state
+
+    ####################################################################################################################
+    # is_complete -- check if syscall has both entry and exit info completed
+    ####################################################################################################################
+    def is_complete(self):
+        all_data = CNT_ENTRY | CNT_EXIT
+        has_only_entry = (self.content & CNT_ENTRY) and (self.mask & EM_no_ret == EM_no_ret)
+        has_only_exit = (self.content & CNT_EXIT) and ((self.ret == 0 and self.name in ("clone", "fork", "vfork")) or
+                                                       self.sc_id == RT_SIGRETURN_SYS_EXIT)
+        return (self.content & all_data == all_data) or has_only_entry or has_only_exit
